# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: hoxbl Spec
  version: 1.0.0
  description: |-
    型安全性強化・API契約強化プロジェクトによるAPI仕様

    Single Source of Truth: Drizzle ORM → Drizzle Zod → Zod → OpenAPI → TypeScript
servers:
  - url: /api
    description: API Server
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT認証トークン
paths:
...
  /tasks:
    get:
      tags:
        - タスク管理
      summary: タスク一覧取得
      description: ログインユーザーのタスク一覧を取得します。フィルタとソートに対応しています。
      security:
        - BearerAuth: []
      parameters:
        - schema:
            type: string
            enum:
              - high
              - medium
              - low
            description: タスクの優先度
            example: high
          required: false
          description: タスクの優先度
          name: priority
          in: query
        - schema:
            type: string
            description: ステータス（カンマ区切りで複数選択可能）
            example: in_progress,in_review
          required: false
          description: ステータス（カンマ区切りで複数選択可能）
          name: status
          in: query
        - schema:
            type: string
            enum:
              - created_at_desc
              - created_at_asc
              - priority_desc
            default: created_at_desc
            description: タスクのソート順
            example: priority_desc
          required: false
          description: タスクのソート順
          name: sort
          in: query
      responses:
        '200':
          description: タスク一覧を取得しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        userId:
                          type: string
                          format: uuid
                        title:
                          type: string
                          minLength: 1
                          maxLength: 100
                        description:
                          type: string
                          nullable: true
                        priority:
                          type: string
                          enum:
                            - high
                            - medium
                            - low
                        status:
                          type: string
                          enum:
                            - not_started
                            - in_progress
                            - in_review
                            - completed
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                      required:
                        - id
                        - userId
                        - title
                        - description
                        - priority
                        - status
                        - createdAt
                        - updatedAt
                required:
                  - success
                  - data
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        anyOf:
                          - type: object
                            additionalProperties:
                              type: string
                          - type: string
                    required:
                      - code
                      - message
                required:
                  - success
                  - error
        '401':
          description: JWT認証失敗
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        anyOf:
                          - type: object
                            additionalProperties:
                              type: string
                          - type: string
                    required:
                      - code
                      - message
                required:
                  - success
                  - error
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        anyOf:
                          - type: object
                            additionalProperties:
                              type: string
                          - type: string
                    required:
                      - code
                      - message
                required:
                  - success
                  - error
    post:
      tags:
        - タスク管理
      summary: タスク作成
      description: 新しいタスクを作成します。
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  nullable: true
                priority:
                  type: string
                  enum:
                    - high
                    - medium
                    - low
                  default: medium
              required:
                - title
      responses:
        '201':
          description: タスクを作成しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      userId:
                        type: string
                        format: uuid
                      title:
                        type: string
                        minLength: 1
                        maxLength: 100
                      description:
                        type: string
                        nullable: true
                      priority:
                        type: string
                        enum:
                          - high
                          - medium
                          - low
                      status:
                        type: string
                        enum:
                          - not_started
                          - in_progress
                          - in_review
                          - completed
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                    required:
                      - id
                      - userId
                      - title
                      - description
                      - priority
                      - status
                      - createdAt
                      - updatedAt
                required:
                  - success
                  - data
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        anyOf:
                          - type: object
                            additionalProperties:
                              type: string
                          - type: string
                    required:
                      - code
                      - message
                required:
                  - success
                  - error
        '401':
          description: JWT認証失敗
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        anyOf:
                          - type: object
                            additionalProperties:
                              type: string
                          - type: string
                    required:
                      - code
                      - message
                required:
                  - success
                  - error
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        anyOf:
                          - type: object
                            additionalProperties:
                              type: string
                          - type: string
                    required:
                      - code
                      - message
                required:
                  - success
                  - error
...